Sort out the types in the traits class

+++ wfmath/const.h
@@ -74,22 +74,22 @@ template<typename FloatType>
 struct numeric_constants
 {
   static const FloatType pi;
-  static const double sqrt_pi;
-  static const double log_pi;
-  static const double sqrt2;
-  static const double sqrt3;
-  static const double log2;
+  static const FloatType sqrt_pi;
+  static const FloatType log_pi;
+  static const FloatType sqrt2;
+  static const FloatType sqrt3;
+  static const FloatType log2;
 };
 
 template<>
 struct numeric_constants<float>
 {
   static const float pi = 3.14159265358979323846264338327950288419716939937508F;
-  static const double sqrt_pi = 1.77245385090551602729816748334114518279754945612237F;
-  static const double log_pi = 1.14472988584940017414342735135305871164729481291530F;
-  static const double sqrt2 = 1.41421356237309504880168872420969807856967187537693F;
-  static const double sqrt3 = 1.73205080756887729352744634150587236694280525381037F;
-  static const double log2 = 0.69314718055994530941723212145817656807550013436025;
+  static const float sqrt_pi = 1.77245385090551602729816748334114518279754945612237F;
+  static const float log_pi = 1.14472988584940017414342735135305871164729481291530F;
+  static const float sqrt2 = 1.41421356237309504880168872420969807856967187537693F;
+  static const float sqrt3 = 1.73205080756887729352744634150587236694280525381037F;
+  static const float log2 = 0.69314718055994530941723212145817656807550013436025F;
 };
 
 /// Determines how close to machine precision the library tries to come.
